services/detect-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
nats-py==2.6.0
pydantic==2.5.3
pydantic-settings==2.1.0
opensearch-py==2.4.2
structlog==24.1.0
python-json-logger==2.0.7
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
